+++ crates/vertigo-cli/src/commons/mod.rs
@@ -1,5 +1,7 @@
 pub mod command;
+mod error_codes;
+pub use error_codes::ErrorCode;
 pub mod models;
-pub mod parse_key_val;
+mod parse_key_val;
 pub use parse_key_val::parse_key_val;
 pub mod spawn;