+++ CHANGES.md
@@ -10,6 +10,7 @@
 ### Changed
 
 * vertigo-cli: Increased statics max-age in Cache-Control header to 1 year to match Google's Lighthouse recommendations
+* vertigo-cli: Improved error messages when building and watching
 
 ## 0.6.1 - 2024-12-18